Replace all anonymous funtions with export ones in order to generate docs
